Configured GitHub repo with proper gitignores, and working USB CDC with ESP8266 Library
Title...

What working here is a git repo that should be recognized by CudeIDE, along with the USB CDC and ESP8266 functionality. "UART Terminal" is the name of the project, the way to get it to work is to connect the esp8266 as normal, and use the micro USB port on the end of the dev board that is opposite the USB debug port we normaly use. This is the "User Target" port, and it's meant to be used for user applications unlike the one that we normally use for debugging, which is wired to an ST-LINK.

The ESP8266 starting sequence works as normal, but with this CDC, we have the added bonus of being able to send commands to the ESP8266 through a terminal like RealTerm or PuTTY, while also receiving the responses that the ESP8266 sends.

Sorry it's so late! I'll get to work testing the full functionality shortly.