Connecting to Kerbal
Connection works, but nothing else. After extensive testing on a source-built plugin, i know that the server is receivng each byte ok, but for some reason it's not being interpreted correctly. Since it works fine with the python client, it might be that either there's something up with our pb defs, or that there there is some unimplemented feature interfering with the data transmission. I can test further with the python client to see whether or not sending the same request generates the same sequence of bytes as the cnano version.